requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8
gunicorn==22.0.0
mysqlclient==2.2.4; sys_platform != "win32"
PyMySQL==1.1.1; sys_platform == "win32"
